Revert "Feature/user registration"
