+++ express_server.js
@@ -1,31 +1,23 @@
-// Import modules
 const express = require("express");
 const app = express();
-const PORT = 8080; // default port 8080
+const PORT = 8081; // default port 8080
 const bodyParser = require("body-parser");
 const cookies = require("cookie-parser");
-const session = require('express-session');
-const flash = require('express-flash');
-const {users, urlDatabase} = require("./database");
-const {generateRandomString, addUser, login} = require("./helper");
 
- 
-
-// Set View Engines
 app.set("view engine", "ejs");
+
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(cookies());
-app.use(session({
-  secret: 'djhxcvxfgshajfgjhgsjhfgsakjeauytsdfy',
-  resave: false,
-  saveUninitialized: true
-  }));
-app.use(flash());
 
+const generateRandomString = function() {
+  return (Math.random()*1e9).toString(36).slice(0,6);
+}
 
+const urlDatabase = {
+  "b2xVn2": "http://www.lighthouselabs.ca",
+  "9sm5xK": "http://www.google.com"
+};
 
-
-// Define get and post methods
 app.get("/", (req, res) => {
   res.redirect("/urls");
 });
@@ -35,20 +27,6 @@ app.get("/urls.json", (req, res) => {
 });
 
 app.get('/urls', (req, res) => {
-  let {user_id} = req.cookies;
-  if (!users[user_id]){
-    user_id = false;
-  }
-  let templateVars = {user_id, urls: urlDatabase};
-  if (user_id && user_id != 'false') {
-    const {id, email, password} = users[user_id]
-    templateVars = {user_id, id, email, password, urls: urlDatabase };
-  }
-  
-  res.render("urls_index", templateVars);
-})
-
-
   const templateVars = {username: req.cookies["username"], urls: urlDatabase };
   res.render("urls_index", templateVars);
 })
@@ -61,67 +39,11 @@ app.get("/urls/new", (req, res) => {
 app.post("/urls", (req, res) => {
   const shortURL = generateRandomString();
   urlDatabase[shortURL] = req.body.longURL;
+  console.log("updated databse", urlDatabase);
   res.redirect(`/urls/${shortURL}`);
 });
 
-app.get("/urls/new", (req, res) => {
-  let {user_id} = req.cookies;
-  if (!users[user_id]){
-    user_id = false;
-  }
-  let templateVars = {user_id};
-  if (user_id && user_id != false) {
-    const {id, email, password} = users[user_id]
-    templateVars = {user_id, id, email, password};
-    res.render("urls_new", templateVars);
-  } else {
-    res.redirect('/urls');    
-  }
-});
-
-app.get('/register', (req, res) => {
-  const {statusCode, statusMessage} = res;
-  const templateVars =  {statusCode, statusMessage};
-  res.render('user_registration', templateVars);
-})
-
-app.post('/register', (req, res) => {
-  const {email, password} = req.body;
-  let user_id = addUser(users, email, password);
-  let {statusCode, statusMessage} = res
-  if ( user_id === "User exists") {
-    statusCode = 400;
-    statusMessage = "User exists"
-    const templateVars =  {statusCode, statusMessage}
-    user_id = false;
-    res.cookie("user_id", user_id);
-    res.render('user_registration', templateVars)
-  } else if (user_id === "Email or Password cannot be empty" ) {
-    statusCode = 400;
-    statusMessage = "Email or Password cannot be empty"
-    const templateVars =  {statusCode, statusMessage}
-    user_id = false;
-    res.cookie("user_id", user_id);
-    res.render('user_registration', templateVars)
-  } else {
-    res.cookie("user_id", user_id);
-    res.redirect("/urls")
-  }
-})
-
 app.get("/urls/:shortURL", (req, res) => {
-  let {user_id} = req.cookies;
-  if (!users[user_id]){
-    user_id = false;
-  }
-  const shortURL = req.params.shortURL;
-  const longURL = urlDatabase[shortURL];
-  let templateVars = {user_id, shortURL, longURL}
-  if (user_id && user_id != false) {
-    const {id, email, password} = users[user_id]
-    templateVars = {user_id, id, email, password, shortURL, longURL};
-  }
-
   const username = req.cookies["username"];
   const shortURL = req.params.shortURL;
   const longURL = urlDatabase[shortURL];
@@ -144,40 +66,6 @@ app.post("/urls/:shortURL/update", (req, res) => {
   urlDatabase[shortURL] = req.body.newURL;
   res.redirect(`/urls/${shortURL}`);
 })
-
-app.get("/login", (req, res) => {
-  const {statusCode, statusMessage} = res;
-  const templateVars =  {statusCode, statusMessage}
-  res.render("user_login", templateVars);
-})
-
-app.post("/login", (req, res) => {
-  const {email, password} = req.body;
-  let {statusCode, statusMessage} = res;
-  const templateVars =  {statusCode, statusMessage}
-  let user_id = login(users, email, password);
-  if (user_id === "Invalid Account") {
-    statusCode = 403;
-    statusMessage = "Invalid Account"
-    const templateVars =  {statusCode, statusMessage}
-    user_id = false;
-    res.cookie("user_id", user_id);
-    res.render("user_login", templateVars)
-  } else if (user_id === "Invalid Password") {
-    statusCode = 403;
-    statusMessage = "Invalid Password"
-    const templateVars =  {statusCode, statusMessage}
-    user_id = false;
-    res.cookie("user_id", user_id);
-    res.render("user_login", templateVars)
-  } else {
-    res.cookie("user_id", user_id);
-    res.redirect('/urls');
-  }
-})
-
-app.post("/logout", (req, res) => {
-  res.clearCookie("user_id");
 
 app.post("/login", (req, res) => {
   res.cookie("username",req.body.username);
