+++ package.json
@@ -20,12 +20,9 @@
   "homepage": "https://github.com/ShuhaoZQGG/tinyapp#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
-    "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.5",
     "ejs": "^3.1.6",
-    "express": "^4.17.1",
-    "express-flash": "git://github.com/RGBboy/express-flash.git",
-    "express-session": "^1.17.2"
+    "express": "^4.17.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.13"
