--- views/user_login.ejs
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <!-- Bootstrap CSS -->
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
-  <title>TinyApp Login</title>
-</head>
-<body>
-  <% if (statusCode !== 200) { %>
-    <div style="margin-top: 100px">
-      <h4 class="text-danger" style = "text-align: center;">Error: <%= statusCode %></h4>
-      <h4 class="text-danger" style = "text-align: center;"><%= statusMessage %></h4>
-    </div>
-  <% } %>
-  <h1 style="text-align: center; margin-top: 100px; -moz-user-select: none;
-  -webkit-user-select: none;
-  -ms-user-select: none;
-  user-select: none;">Welcome to Shortening URL TinyApp</h1>
-  <form method = 'POST' action = '/login' style="text-align: center; -moz-user-select: none;
-  -webkit-user-select: none;
-  -ms-user-select: none;
-  user-select: none;">
-    Email: <input name = "email" type = "email" style="height: 40px;  margin-top: 40px;">
-    <br>
-    Password: <input name = "password" type = "password" style="height: 40px; margin-top: 40px; margin-right: 25px;">
-    <br>
-    <button type ="submit" class="btn btn-primary" style = "margin-top: 40px;">Log In</button>
-  </form>
-  <form method="GET" action="/register" style="text-align: center; margin-top: 20px;-moz-user-select: none;
-  -webkit-user-select: none;
-  -ms-user-select: none;
-  user-select: none;">
-    <button type ="submit" class="btn btn-warning">I don't have an account, Register</button>
-  </form>
-  <form method="GET" action="/urls" style="text-align: center; margin-top: 60px;-moz-user-select: none;
-  -webkit-user-select: none;
-  -ms-user-select: none;
-  user-select: none;">
-    <button type ="submit" class="btn btn-secondary">Continue as a Guest</button>
-  </form>
-    <!-- Bootstrap JS -->
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script> 
-</body>
-</html>
